fix(create-routecraft): remove useSrcDir option and fix index.ts path bug
The useSrcDir option caused index.ts to always be placed at the project
root while package.json start script pointed to src/index.ts, making
generated projects fail to start. Remove the option entirely to simplify
the scaffolder -- projects always use a flat structure.

Also removes the tsc build script from the template since craft run
executes TypeScript directly, and extracts lib.ts for testability.

Rewrites the test suite with 31 tests: unit tests for pure functions,
scaffolding tests for all option combinations, and integration tests
that install deps, typecheck, and run a generated project end-to-end.
