+++ packages/create-routecraft/templates/base/index-with-example.ts
@@ -1,5 +1,5 @@
 export { craftConfig } from "./craft.config.js";
-import helloWorldRoute from "CAPABILITIES_IMPORT_PATH/hello-world.js";
+import helloWorldRoute from "./capabilities/hello-world.js";
 
 // Export all routes as default for craft run
 export default [helloWorldRoute];
