+++ packages/create-routecraft/templates/base/tsconfig.json
@@ -3,7 +3,6 @@
     "target": "ES2022",
     "module": "NodeNext",
     "moduleResolution": "NodeNext",
-    "outDir": "./dist",
     "rootDir": ".",
     "allowSyntheticDefaultImports": true,
     "esModuleInterop": true,
